% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_urls.R
\name{config_urls}
\alias{config_urls}
\title{Split a string}
\usage{
config_urls()
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{What to split on.}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
